design-projects/distributed-tracing-with-open-telemetry/processor/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg2-binary
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp-proto-http